backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
sentence-transformers>=2.3.0
transformers>=4.35.0,<5.0.0
torch>=2.1.1
# Note: For GPU support, install torch with CUDA:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# Or use: pip install torch --index-url https://download.pytorch.org/whl/cu121 for CUDA 12.1
huggingface_hub>=0.20.0
numpy==1.24.3
scikit-learn==1.3.2
chromadb>=0.5.0
faiss-cpu==1.7.4
qdrant-client>=1.7.0
langchain==0.0.350
langchain-community==0.0.10
pydantic==2.5.0
pydantic[email]==2.5.0
email-validator==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1

